Move to query/ir, fix u_set vs set bug
