+++ query/ir/CMakeLists.txt
@@ -3,3 +3,4 @@ add_subdirectory(common)
 add_subdirectory(as)
 add_subdirectory(interpreter)
 add_subdirectory(lowering)
+add_subdirectory(frontend)
